Widen the transform update id so collider caches cannot miss a wrap (#204)
Transform::m_updateID was uint8_t, and BoxCollider's cached mesh id and
BoundingBox::lastUpdateID compared against it to decide whether cached
collision geometry was still valid. The counter wrapped every 256
updates, so a collider queried only occasionally (a raycast target)
could sit exactly 256 updates stale and skip a rebuild, serving
outdated collision geometry.

Widen the counter to uint64_t everywhere it is stored or compared:
Transform::m_updateID and getUpdateID(), BoxCollider's cached
transform id, and BoundingBox::lastUpdateID. A 64-bit counter cannot
realistically wrap.

The update id is never packed or serialized (Transform::pack/unpack
and serialize/loadFromJSON only carry position/scale/rotation), so
this does not change the wire or save format.

Separately: BoundingBox::lastUpdateID defaulting to 0 while
Transform::m_updateID defaults to 1 no longer matters for correctness
- Collider::m_boundingBoxDirty and BoxCollider::m_meshDirty both
default to true, so the first bounding box/mesh computation is forced
regardless of whether the two defaults happen to differ.

Added a test to ColliderWorldTransformTest.cpp that populates a box
collider's mesh cache, drives the transform through exactly 256
updates ending at a position the cache has never seen, and asserts
the support point reflects the new position. It fails on pre-fix code
because the id wraps back to the cached value after 256 uint8_t
increments.

+++ source/libs/data/objects/components/Transform.cpp
@@ -20,7 +20,7 @@ Transform::Transform(const glm::vec3& position, const glm::vec3& scale, const gl
   loadVariable(m_rotation);
 }
 
-uint8_t Transform::getUpdateID() const
+uint64_t Transform::getUpdateID() const
 {
   return m_updateID;
 }
